Remove openssl #192
